Update azure_almalinux_8.py
Adding nct4v3, along with the previous addition of a100v4. Using either 'gpu' or 'nct4v3' for the partition name does not make a different, functionally. This is because our config file is a hack of how reframe config files normally work. The system names in our config files are actually the partition names on a VM cluster. Therefore, we run tests based on valid systems instead of valid partitions within a system (ex: hbv3 is really a partition instead of a system - therefore, having no partitions within it). In other words, the partition names are not relevant.